added if exist condition fpr down las_use_at

+++ database/migrations/2022_02_07_105428_add_last_use_to_addresses_table.php
@@ -26,9 +26,10 @@ class AddLastUseToAddressesTable extends Migration
      */
     public function down()
     {
-        Schema::table('addresses', function (Blueprint $table) {
-            //
-            $table->dropColumn('last_use_at');
-        });
+        if (Schema::hasColumn('addresses', 'last_use_at')){
+            Schema::table('addresses', function (Blueprint $table) { 
+                $table->dropColumn('last_use_at');
+            });
+        }
     }
 }
